Sesión 27 de noviembre de 2020

+++ borrar.php
@@ -1,8 +1,15 @@
 <?php
+session_start();
+
+if (isset($_SESSION['autenticado']) && $_SESSION['autenticado']) {
     $nombre = $_GET['nombre'];
-    
+
     unlink("posts/$nombre");
-    
-    header("Location: index.php");
-        
-?>
+    $_SESSION['info'] = 'Post Borrado';
+    unset($_SESSION['error']);
+}else{
+    $_SESSION['info'] = 'Acción no permitida';
+    unset($_SESSION['error']);
+}
+
+header("Location: index.php");
